requirements.txt
# Focus Server Automation Framework - Dependencies
# ================================================

# Core Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-timeout>=2.2.0
pytest-mock>=3.12.0
pytest-html>=4.1.1
pytest-cov>=4.1.0
pytest-json-report>=1.5.0
pytest-xdist>=3.3.1  # For parallel execution

# HTTP and API Testing
requests>=2.31.0
httpx>=0.24.1  # Modern HTTP client
beautifulsoup4>=4.12.0  # For API Healing Framework (Swagger parsing)

# Browser Automation and UI Testing
playwright>=1.40.0  # AI-powered UI testing with self-healing selectors
pytest-playwright>=0.4.3  # Pytest plugin for Playwright

# Infrastructure Management
kubernetes>=28.1.0
pymongo>=4.6.0
paramiko>=3.3.1
pika>=1.3.0  # RabbitMQ client

# Data Processing and Validation
pydantic>=2.4.0
pydantic-settings>=2.0.0
orjson>=3.9.10

# Configuration Management
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging and Monitoring
structlog>=23.2.0
colorlog>=6.7.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
psutil>=5.9.6

# Development Tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Security
cryptography>=41.0.7
bandit>=1.7.5

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Reporting
allure-pytest>=2.13.0
jinja2>=3.1.2

# Async Support
asyncio-mqtt>=0.16.1
aiofiles>=23.2.0

# Network Testing
netaddr>=0.10.1
# socket-client>=1.0.0  # This package doesn't exist, removing

# Jira Integration
jira>=3.5.0  # Jira API client for Python
pytest-xray>=3.0.0  # Optional: pytest plugin for Xray integration (makes marker handling easier)

# File System Operations
# pathlib2>=2.3.7  # Not needed in Python 3.4+, using built-in pathlib

# UUID Generation (built-in, but explicit)
# uuid - built-in Python module

# Threading Utilities
# concurrent.futures - built-in Python module

# Subprocess Management (built-in)
# subprocess - built-in Python module

# JSON Handling (built-in, but orjson for performance)
# json - built-in Python module

# Time utilities (built-in)
# time - built-in Python module
# datetime - built-in Python module

# Environment variable management
# os - built-in Python module

# File system operations
# pathlib - built-in Python module

# Regular expressions
# re - built-in Python module

# System information
# platform - built-in Python module
# sys - built-in Python module

# Mathematical operations
# math - built-in Python module

# Data structures
# collections - built-in Python module
# typing - built-in Python module
# dataclasses - built-in Python module
# enum - built-in Python module